postgresql_16: build with meson instead of autotools
This build the latest postgresql version with meson instead of autotools
to support building in pkgsStatic in the next commit.

TODO: The meson build does not produce bitcode (.bc) files when JIT is
enabled - this is not implemented upstream, yet.  This probably means
that JIT is currently not working, but not tested, yet.