pkgsStatic.postgresql_16: fix build
This disables shared libraries for a meson-based pkgsStatic.postgresql
build.  This is mostly useful for a static libpq, but a static server
binary is also built.  In-core and contrib extensions are also built,
but not all of them work correctly.  However, this could still be
helpful to allow running tests in the static overlay with a running
postgres instance - most of those tests, should not need the full
functionality of PostgreSQL.

A nice side-effect of the patch is, that static libraries are not built
anymore for the default build, so we don't need to remove them manually
anmyore.

TODO:
- The various binaries in the default output are still dynamically
linked to libc (musl). This is probably beause otherwise loading
extensions would not be supported. Maybe we can limit the postgres
binary to that and keep all other binaries strictly static, though, if
we can pass "-Wl,-Bstatic" to all other binaries.
- Figure out how to make the failing tests pass and the server fully
functional, if at all possible.