postgresql: fix regress tests after tzdata update
This was discussed upstream here:
https://postgr.es/m/0a997455-5aba-4cf2-a354-d26d8bcbfae6%40technowledgy.de

This will be in the next minor releases, until then apply the patch.
Applying the patch from github via fetchurl is not possible, because the
upstream patch includes some doc changes, which cause our build to fail.
We could fix that by adding new dependencies to rebuild the docs... but
it's easier to just keep the patch in nixpkgs for 1-2 months.

TLDR: The patches have been slightly modified from upstream, removing
the docs changes.