workflows/get-merge-commit: inline get-merge-commit.sh script
This removes a bit of code, because checking arguments and separating
output streams to stderr and stdout is not required anymore. The reason
this was a separate shell script was, that this would be included in
multiple workflows separately. But a while ago this had been changed to
a re-usable workflow, so we can just as well inline the script.

This will later have the benefit that we don't need a separate checkout
for base/ci in this step, because the script is already part of the
workflow file. This can not be done, yet, because
ci/supportedSystems.json is still required.