workflows: checkout nixpkgs in get-merge-commit action
This makes checking out the nixpkgs repo even more consistent and also
fixes a bug introduced lately: After enabling CI to run on merge
conflicted PRs, the logic in eval to fetch the target commit doesn't
work anymore in PRs with more than one commit. That's because the
"target" commit is then not the direct parent of the "merged" commit,
since the base and the head of the branch have commits between them.

By fetching both commits explicitly we do the right thing in all cases.