workflows/eval: skip on ready_for_review
We don't need to run full eval when undrafting a PR. We already have an
eval result, so we can use that to do the maintainer pings.

We need to wait for eval to finish first, but we're already half there
because of how we're waiting for the artifact to appear. Since the
ready_for_review case is triggered in a different workflow run, we'll
need to fetch the ID of the relevant Eval workflow first, though.