workflows/check-cherry-picks: reduce checkout time
Using a `tree:0` filter instead of `blob:none` reduces the checkout time
from over 3 minutes to about 45 seconds. The required trees/blobs will
then be fetched on-demand.

This on-demand fetching creates additional output for `git range-diff`,
so we `--quiet` this to only get the result we actually want to see.