workflows/lib-tests: use nixpks CI cachix cache
Instead of depending the workflow on changed paths, we use cachix to
cache the results. No changes to any of those folders means the job will
be done very quickly as well.

Yes, this will still be a loss, because the job will be run *a lot more
often*. But running those jobs unconditionally is a pre-requisite to
enable required status checks / workflows eventually, so let's work
towards that.