workflows/check-shell: use nixpkgs CI cachix cache
The check-shell currently needs to rebuild treefmt on all 4 systems each
time. Since this depends on the pinned nixpkgs and the pinned
treefmt-nix, this is a prime candidate for caching.

With a cached build, this should be very fast, so we enable building on
all PRs to unblock *required status checks / workflows* in the future.