workflows/manual-{nixos,nixpkgs}: run without condition on changed paths
To enable *required status checks / workflows* in the future, we'd like
to run all workflows unconditionally. Since those two workflows are
already using cachix, the additional runs will be very cheap. Yes, we'll
run additional jobs, but that will be temporary only, see next commits.

The upside is, that we're not going to accidentally miss some of the
paths that would cause rebuilds as we did in the past.