workflows/lint: run all the static lints in same workflow
They still run in separate jobs, but they give us a nice consistent
appearance in the check list: All of them are starting with "Lint /
...".

Most importantly, the fewer workflows we have, the easier it is to set
them up as "required workflows" in branch protection rules. That's also
why we take care to have a really short and simple name like "Lint" that
is not going to change anymore.