workflows/check: rename from "No channel PR"
This introduces "check" as another category of jobs suitable for the
"Require status checks to pass" feature.

Checks as a category would include everything that is done before even
looking at the code: Right branch? Commit messages? Cherry-picked
correctly?

(the latter is moved later, because it first needs to heavily changed
elsewhere)