workflows: use consistent naming with the 4 main workflows
The idea is that the four main workflows have no extra "name" attributes
and use their file / attribute names. This is to make it harder to
accidentally change a name, for example "when just fixing a typo
quickly" - and then running the risk of running the branch protection
rules out of sync. This would lead to all merges being blocked, so we'd
really want to minimize those chances.

By removing the "name:" attribute from all workflows and the jobs
(except matrix), we're growing a habit of never having them, thus making
it less likely to accidentally introduce them in the critical workflows.

Also removed the -v2 suffix from codeowners, but this needs a rename to
owners, because the "codeowners" workflow has been explicitly disabled
for security reasons.