workflows/created: init
This new workflow only runs once after creation of a pull request to add
a footer. This will make the pull request template a bit shorter to read
through - which is OK, because this information *is not relevant at all
during creation of the pull request*. At the same time it ensures that
this note is present on *all* pull requests, even those where the author
decides to use no template, on automatically created backports, reverts,
and so on.